17.1 Sweetalert2

+++ resources/views/users/index.blade.php
@@ -36,16 +36,34 @@
 @section('javascript')
     $(function() {
         $('.delete').click(function (){
-            $.ajax({
-                method: "DELETE",
-                url: $(this).data("id")
+
+            Swal.fire({
+                title: 'Na pewno chcesz usunąć rekord?',
+                text: "Już tego nie cofniesz!",
+                icon: 'warning',
+                showCancelButton: true,
+                confirmButtonColor: '#3085d6',
+                cancelButtonColor: '#d33',
+                confirmButtonText: 'Tak, usuń',
+                cancelButtonText: 'Nie usuwaj'
+            }).then((result) => {
+                if (result.isConfirmed) {
+                    $.ajax({
+                        method: "DELETE",
+                        url: $(this).data("id")
+                    })
+                    .done(function( response ) {
+                        window.location.reload()
+                    })
+                    .fail(function( response ) {
+                        Swal.fire({
+                            icon: 'error',
+                            title: 'Oj...!',
+                            text: 'Błąd. Coś poszło nie tak.'
+                        })
+                    });
+                }
             })
-            .done(function( response ) {
-                window.location.reload();
-            })
-            .fail(function( response ) {
-                alert( "BŁĄD" );
-            });
+        });
     });
-});
 @endsection
